Flytter noen filer tilbake for en mer leselig git-diff
Lavthengende frukt for leselighet å skille dette ut til en egen PR
