+++ src/sider/01-personalia/Kontonr.tsx
@@ -2,12 +2,12 @@ import * as React from "react";
 import {startTransition, useState} from "react";
 import {KontonrShow} from "./KontonrShow.tsx";
 import {KontonrEdit} from "./KontonrEdit.tsx";
-import {Systeminfo} from "../../../lib/components/systeminfo/Systeminfo.tsx";
+import {Systeminfo} from "../../lib/components/systeminfo/Systeminfo.tsx";
 import {Loader} from "@navikt/ds-react";
-import {useKontonummer} from "../../../lib/hooks/data/useKontonummer.ts";
-import {PersonaliaEditKnapp} from "../PersonaliaEditKnapp.tsx";
+import {useKontonummer} from "../../lib/hooks/data/useKontonummer.ts";
+import {PersonaliaEditKnapp} from "./PersonaliaEditKnapp.tsx";
 
-export const Kontoinformasjon = () => {
+export const Kontonr = () => {
     const [editMode, setEditMode] = useState<boolean>(false);
     const {kontoinformasjon, updateKontoInformasjon, isLoading, isMutating} = useKontonummer();
