+++ src/sider/01-personalia/KontonummerFormSchema.ts
