+++ src/sider/08-vedlegg/form/AvdragRenter.tsx
@@ -1,11 +1,11 @@
-import {AvdragRenterDto, BoliglanInput} from "../../../../../generated/new/model";
+import {AvdragRenterDto, BoliglanInput} from "../../../generated/new/model";
 import {useForm} from "react-hook-form";
 import React from "react";
 import {zodResolver} from "@hookform/resolvers/zod";
-import {avdragRenterFormToBoliglanInput} from "./avdragRenterFormToBoliglanInput.ts";
-import {AvdragRenterFormSchema, AvdragRenterFormValues} from "./AvdragRenterFormSchema.ts";
 import {AvdragRenterInputList} from "./AvdragRenterInputList.tsx";
-import {DokumentasjonTypesForVariant} from "../../../../../lib/opplysninger.ts";
+import {DokumentasjonTypesForVariant} from "../../../lib/opplysninger.ts";
+import {AvdragRenterFormSchema, AvdragRenterFormValues} from "./lib/formSchemas.ts";
+import {avdragRenterFormToBoliglanInput} from "./lib/formToInputMappers.ts";
 
 const AvdragRenter = ({
     opplysningstype,
