+++ src/sider/08-vedlegg/form/BruttoNetto.tsx
@@ -1,12 +1,12 @@
-import {DokumentasjonDtoType, LonnsInntektDto, type LonnsInput} from "../../../../../generated/new/model";
+import {DokumentasjonDtoType, LonnsInntektDto, type LonnsInput} from "../../../generated/new/model";
 import {useForm} from "react-hook-form";
 import React from "react";
 import {zodResolver} from "@hookform/resolvers/zod";
 import {z} from "zod";
-import {BruttoNettoFormSchema} from "./BruttoNettoFormSchema.ts";
-import {bruttoNettoFormToLonnsInput} from "./bruttoNettoFormToLonnsInput.ts";
-import {OpplysningBelopInput} from "../../common/OpplysningBelopInput.tsx";
-import {useDokumentasjonTekster} from "../../../../../lib/hooks/dokumentasjon/useDokumentasjonTekster.ts";
+import {OpplysningBelopInput} from "./components/OpplysningBelopInput.tsx";
+import {useDokumentasjonTekster} from "../../../lib/hooks/dokumentasjon/useDokumentasjonTekster.ts";
+import {BruttoNettoFormSchema} from "./lib/formSchemas.ts";
+import {bruttoNettoFormToLonnsInput} from "./lib/formToInputMappers.ts";
 
 export type BruttoNettoFormValues = z.infer<typeof BruttoNettoFormSchema>;
 const BruttoNetto = ({
