+++ src/sider/08-vedlegg/form/components/FormSwitch.tsx
@@ -1,15 +1,15 @@
-import useOkonomiskOpplysningMutation from "../../../lib/hooks/dokumentasjon/useOkonomiskOpplysningMutation.ts";
-import {DokumentasjonTypesForVariant, opplysningSpec} from "../../../lib/opplysninger.ts";
+import useOkonomiskOpplysningMutation from "../../../../lib/hooks/dokumentasjon/useOkonomiskOpplysningMutation.ts";
+import {DokumentasjonTypesForVariant, opplysningSpec} from "../../../../lib/opplysninger.ts";
 import {Loader} from "@navikt/ds-react";
-import AvdragRenter from "./variants/avdragRenter/AvdragRenter.tsx";
-import BelopBeskrivelse from "./variants/belopBeskrivelse/BelopBeskrivelse.tsx";
-import BruttoNetto from "./variants/bruttoNetto/BruttoNetto.tsx";
-import BelopEn from "./variants/belopEn/BelopEn.tsx";
+import AvdragRenter from "../AvdragRenter.tsx";
+import BelopBeskrivelse from "../BelopBeskrivelse.tsx";
+import BruttoNetto from "../BruttoNetto.tsx";
+import BelopEn from "../BelopEn.tsx";
 import React from "react";
-import {DokumentasjonDtoType} from "../../../generated/new/model/dokumentasjonDtoType.ts";
-import {AvdragRenterDto} from "../../../generated/new/model/avdragRenterDto.ts";
-import {BelopDto} from "../../../generated/new/model/belopDto.ts";
-import {LonnsInntektDto} from "../../../generated/new/model/lonnsInntektDto.ts";
+import {DokumentasjonDtoType} from "../../../../generated/new/model/dokumentasjonDtoType.ts";
+import {AvdragRenterDto} from "../../../../generated/new/model/avdragRenterDto.ts";
+import {BelopDto} from "../../../../generated/new/model/belopDto.ts";
+import {LonnsInntektDto} from "../../../../generated/new/model/lonnsInntektDto.ts";
 
 export const FormSwitch = ({opplysningstype}: {opplysningstype: DokumentasjonDtoType}) => {
     const {updateOkonomiskOpplysning, opplysning, isLoading} = useOkonomiskOpplysningMutation(opplysningstype);
