+++ src/sider/08-vedlegg/form/components/OpplysningTextInput.tsx
@@ -1,10 +1,10 @@
 import {TextField, TextFieldProps} from "@navikt/ds-react";
 import {useController, UseControllerProps} from "react-hook-form";
-import {BelopBeskrivelseFormValues} from "../variants/belopBeskrivelse/BelopBeskrivelseFormSchema.ts";
 import cx from "classnames";
 import React from "react";
 import {DigisosLanguageKey} from "../../../../lib/i18n/common.ts";
 import {useTranslation} from "react-i18next";
+import {BelopBeskrivelseFormValues} from "../lib/formSchemas.ts";
 
 export const OpplysningTextInput = ({
     label,
