--- src/sider/08-vedlegg/form/variants/belopEn/BelopEnSchema.ts
@@ -1,8 +1,0 @@
-import {z} from "zod";
-import {BelopSchema} from "../../common/BelopSchema.ts";
-
-export const BelopEnSchema = z.object({
-    belop: BelopSchema.optional(),
-});
-
-export type BelopEnFormValues = z.infer<typeof BelopEnSchema>;
