--- src/sider/08-vedlegg/form/variants/bruttoNetto/BruttoNettoFormSchema.ts
@@ -1,7 +1,0 @@
-import {z} from "zod";
-import {BelopSchema} from "../../common/BelopSchema.ts";
-
-export const BruttoNettoFormSchema = z.object({
-    brutto: BelopSchema.optional(),
-    netto: BelopSchema.optional(),
-});
